+++ appseed/ca/sockets/bsd/sockets_link_in_socket.cpp
@@ -52,7 +52,7 @@ namespace sockets
    void link_in_socket::link_write(void * p, size_t n)
    {
       
-      m_in->SendBuf((const char *) p, n);
+      m_in->SendBuf((const char *) p, (int32_t) n);
 
    }
