+++ appseed/ca/sockets/bsd/sockets_link_out_socket.cpp
@@ -48,7 +48,7 @@ namespace sockets
 
    void link_out_socket::link_write(void * p, size_t size)
    {
-      m_out->SendBuf((const char *) p, size);
+      m_out->SendBuf((const char *) p, (int32_t) size);
    }
